# basename_natural: Improve performance