--- a/tests/ranger/container/test_fsobject.py
+++ b/tests/ranger/container/test_fsobject.py
@@ -30,7 +30,7 @@ def test_basename_natural1():
             "hello",
             "hello1", "hello2",
             "hello11", "hello12",
-            "hello100", "hello101", "hello111", "hello112",
+            "hello100", "hello101", "hello111", "hello112"
         )
     ]
     assert fsos == sorted(fsos[::-1], key=operator.attrgetter("basename_natural"))
@@ -42,11 +42,12 @@ def test_basename_natural2():
     fsos = [
         create_filesystem_object(path)
         for path in (
-            "hello", "hello.txt",
-            "hello0.txt", "hello1.txt", "hello2.txt", "hello3.txt"
-            "hello10.txt", "hello11.txt", "hello12.txt", "hello13.txt"
-            "hello100.txt", "hello101.txt", "hello102.txt", "hello103.txt"
-            "hello110.txt", "hello111.txt", "hello112.txt", "hello113.txt"
+            "hello",
+            "hello0.txt", "hello1.txt", "hello2.txt", "hello3.txt",
+            "hello10.txt", "hello11.txt", "hello12.txt", "hello13.txt",
+            "hello100.txt", "hello101.txt", "hello102.txt", "hello103.txt",
+            "hello110.txt", "hello111.txt", "hello112.txt", "hello113.txt",
+            "hello.txt"
         )
     ]
     assert fsos == sorted(fsos[::-1], key=operator.attrgetter("basename_natural"))