@toonn 2024-06-11
This looks like a pretty serious improvement.

I am a bit wary of affecting the collation order in somewhat unexpected ways though. My understanding of the current implementation is that it's faithful to [the definition Wikipedia gives](https://en.wikipedia.org/wiki/Natural_sort_order):

> In computing, natural sort order (or natural sorting) is the ordering of strings in alphabetical order, except that multi-digit numbers are treated atomically, i.e., as if they were a single character.

We put numbers into a tuple with `'0'` as the first element, so any other character will compare to the number as if it's a single character. The only case where we do numeric comparison is if both tuples contain a number and therefore have the same first element.

However, further reading of some links on that article, like [Dave Koelle's blog](https://web.archive.org/web/20210803201519/http://www.davekoelle.com/alphanum.html) seem to show an algorithm that's more in line with what you have here. Though [Martin Pool makes a good point here](https://sourcefrog.net/projects/natsort/) about leading zeroes. `1.002 < 1.01` at least to our  "natural" thinking if we interpret them as fractional numbers but both the current implementation and the new one would compare `[('0', 1), ('.', 0), ('0', 2)] > [('0', 1), ('.', 0), ('0', 1)]`.

I'm not sure about fixing this though because what about scientific notation or dates, dates have a natural sort order but they come in many formats. And there's other cases where natural sorting runs afoul of expectations, like hashes, I don't actually want coincidental stretches of digits affecting their sort order if I'm being honest.

After all of this I figured we should just defer to the wisdom of our elders and go with what `ls` does. GNU ls has the `-v` flag, which is equivalent to `--sort=version` and intends to sort version numbers naturally. However, that seems to lead to yet another order? `a   a1   ab  'a 2'  'a b'  'a #1'   a#`

In summary, I'm open to this change but want to check in with the community before changing the behavior. It wouldn't be hard to introduce a setting to flip between the old and new orders or introduce a new "naturalish" sort order.