@mingmingrr 2025-09-14
I made some additional changes that should keep the previous sort behavior, at the cost of a ~25% performance penalty:
```
Original:
55140  4.119  4.634  fsobject.py:167(basename_natural_lower)
Sort behavior changed:
55140  0.998  1.310  fsobject.py:167(basename_natural_lower)
Sort behavior unchanged:
55140  1.238  1.664  fsobject.py:168(basename_natural_lower)
```